Modularize code: split index.ts into firebase.ts and reverseproxy.ts
So the ReverseProxy logic reusable, rather than only exposing a firebase-specific reverse proxy.

This makes it easier to build workers that reverse proxy to firebase and non-firebase endpoints
